backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI for GPT-4o Vision
openai==1.3.5

# Playwright for browser automation
playwright==1.40.0

# OCR fallback
pytesseract==0.3.10
Pillow==10.1.0
easyocr==1.7.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# HTML Parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Image processing
opencv-python==4.8.1.78
numpy==1.24.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Resume parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

